cleaning CompileSwiftTask
moving all command for updating the manifest to CreatePackageSwift
